+++ CHANGELOG.md
@@ -6,10 +6,10 @@
 
 ## [0.2.0]
 ### Changed
-- Changed the file chooser dialog in the 'Open Index' and 'Create Index' dialogs to an IntelliJ one to provide
+- Changed the file chooser dialog in the *Open Index*, *Create Index* dialogs and the *Custom Analyzer* panel to an IntelliJ one to provide
 a smoother user experience.
-- Replaced the previous GIF loading icon to a sharper SVG one. 
+- Replaced the previous GIF loading icon to a sharper SVG one.
 
 ## [0.1.0]
 ### Added
-- Integrated Lucene's Luke application into this IDE plugin. Code changes from Luke are applied up to version 9.10.0. 
+- Integrated Lucene's Luke application into this IDE plugin. Code changes from Luke are applied up to version 9.10.0.